feat: clean up duplicate function
