+++ mirascope/core/base/messages.py
@@ -5,7 +5,7 @@ from collections.abc import Sequence
 from . import AudioPart, BaseMessageParam, ImagePart, TextPart
 from ._utils._convert_messages_to_message_params import (
     Image,
-    get_content_from_message,
+    convert_message_content_to_message_param_content,
 )
 
 
@@ -23,7 +23,10 @@ class Messages:
             content: str
             | Sequence[str | Image.Image | TextPart | ImagePart | AudioPart],
         ) -> None:
-            super().__init__(content=get_content_from_message(content), role="system")
+            super().__init__(
+                content=convert_message_content_to_message_param_content(content),
+                role="system",
+            )
 
     class User(BaseMessageParam):
         def __init__(
@@ -31,7 +34,10 @@ class Messages:
             content: str
             | Sequence[str | Image.Image | TextPart | ImagePart | AudioPart],
         ) -> None:
-            super().__init__(content=get_content_from_message(content), role="user")
+            super().__init__(
+                content=convert_message_content_to_message_param_content(content),
+                role="user",
+            )
 
     class Assistant(BaseMessageParam):
         def __init__(
@@ -40,6 +46,6 @@ class Messages:
             | Sequence[str | Image.Image | TextPart | ImagePart | AudioPart],
         ) -> None:
             super().__init__(
-                content=get_content_from_message(content),
+                content=convert_message_content_to_message_param_content(content),
                 role="assistant",
             )
