+++ mirascope/core/base/prompt.py
@@ -1,14 +1,13 @@
 """The `BasePrompt` class for better prompt engineering."""
 
 import types
-from collections.abc import AsyncIterable, Awaitable, Callable, Iterable, Sequence
+from collections.abc import AsyncIterable, Awaitable, Callable, Iterable
 from functools import reduce, wraps
 from typing import (
     Any,
     ClassVar,
     ParamSpec,
     Protocol,
-    TypeAlias,
     TypeVar,
     overload,
 )
@@ -24,14 +23,11 @@ from ._utils import (
     get_prompt_template,
     parse_prompt_messages,
 )
-from ._utils._convert_messages_to_message_params import (
-    convert_messages_to_message_params,
-)
+from ._utils._messages_decorator import MessagesDecorator, _messages_decorator
 from ._utils._protocols import fn_is_async
 from .call_response import BaseCallResponse
 from .dynamic_config import BaseDynamicConfig
 from .message_param import BaseMessageParam
-from .messages import Messages
 from .metadata import Metadata
 from .stream import BaseStream
 
@@ -304,17 +300,6 @@ class BasePrompt(BaseModel):
 
 
 _BasePromptT = TypeVar("_BasePromptT", bound=BasePrompt)
-_MessageFuncReturnTypes: TypeAlias = BaseDynamicConfig | Messages.Type
-_MessageFuncReturnT = TypeVar(
-    "_MessageFuncReturnT", bound=_MessageFuncReturnTypes, contravariant=True
-)
-
-
-def _is_messages_type(value: object) -> TypeIs[Messages.Type]:
-    return isinstance(
-        value,
-        str | Sequence | list | BaseMessageParam,
-    )
 
 
 class PromptDecorator(Protocol):
@@ -337,8 +322,8 @@ class PromptDecorator(Protocol):
         | Callable[_P, BaseDynamicConfig]
         | Callable[_P, Awaitable[BaseDynamicConfig]],
     ) -> (
-        Callable[_P, str | list[BaseMessageParam]]
-        | Callable[_P, Awaitable[str | list[BaseMessageParam]]]
+        Callable[_P, list[BaseMessageParam]]
+        | Callable[_P, Awaitable[list[BaseMessageParam]]]
         | type[_BasePromptT]
     ): ...
 
@@ -351,79 +336,6 @@ def _is_base_dynamic_config_function(
     Callable[_P, BaseDynamicConfig] | Callable[_P, Awaitable[BaseDynamicConfig]]
 ]:
     return isinstance(prompt, types.FunctionType)
-
-
-MessagesSyncFunction: TypeAlias = Callable[_P, _MessageFuncReturnT]
-MessagesAsyncFunction: TypeAlias = Callable[_P, Awaitable[_MessageFuncReturnT]]
-
-
-class MessagesDecorator(Protocol):
-    @overload
-    def __call__(
-        self,
-        messages_fn: MessagesSyncFunction[_P, _MessageFuncReturnT],
-    ) -> Callable[_P, list[BaseMessageParam] | BaseDynamicConfig]: ...
-
-    @overload
-    def __call__(
-        self,
-        messages_fn: MessagesAsyncFunction[_P, _MessageFuncReturnT],
-    ) -> Callable[_P, Awaitable[list[BaseMessageParam] | BaseDynamicConfig]]: ...
-
-    def __call__(
-        self,
-        messages_fn: MessagesSyncFunction[_P, _MessageFuncReturnT]
-        | MessagesAsyncFunction[_P, _MessageFuncReturnT],
-    ) -> (
-        Callable[_P, list[BaseMessageParam] | BaseDynamicConfig]
-        | Callable[_P, Awaitable[list[BaseMessageParam] | BaseDynamicConfig]]
-    ): ...
-
-
-def _messages_decorator() -> MessagesDecorator:
-    @overload
-    def inner(
-        messages_fn: MessagesAsyncFunction[_P, _MessageFuncReturnT],
-    ) -> Callable[_P, Awaitable[list[BaseMessageParam] | BaseDynamicConfig]]: ...
-
-    @overload
-    def inner(
-        messages_fn: MessagesSyncFunction[_P, _MessageFuncReturnT],
-    ) -> Callable[_P, list[BaseMessageParam] | BaseDynamicConfig]: ...
-
-    def inner(
-        messages_fn: MessagesSyncFunction[_P, _MessageFuncReturnT]
-        | MessagesAsyncFunction[_P, _MessageFuncReturnT],
-    ) -> (
-        Callable[_P, Awaitable[list[BaseMessageParam] | BaseDynamicConfig]]
-        | Callable[_P, list[BaseMessageParam] | BaseDynamicConfig]
-    ):
-        if fn_is_async(messages_fn):
-
-            @wraps(messages_fn)
-            async def get_base_message_params_async(
-                *args: _P.args, **kwargs: _P.kwargs
-            ) -> list[BaseMessageParam] | BaseDynamicConfig:
-                raw_messages = await messages_fn(*args, **kwargs)
-                if _is_messages_type(raw_messages):
-                    return convert_messages_to_message_params(raw_messages)
-                return raw_messages
-
-            return get_base_message_params_async
-        else:
-
-            @wraps(messages_fn)
-            def get_base_message_params(
-                *args: _P.args, **kwargs: _P.kwargs
-            ) -> list[BaseMessageParam] | BaseDynamicConfig:
-                raw_messages = messages_fn(*args, **kwargs)
-                if _is_messages_type(raw_messages):
-                    return convert_messages_to_message_params(raw_messages)
-                return raw_messages
-
-            return get_base_message_params
-
-    return inner
 
 
 @overload
