+++ tests/core/base/_utils/test_convert_messages_to_message_params.py
@@ -6,8 +6,8 @@ from mirascope.core import BaseMessageParam
 from mirascope.core.base import AudioPart, ImagePart, TextPart
 from mirascope.core.base._utils._convert_messages_to_message_params import (
     _convert_message_sequence_part_to_content_part,
-    _convert_message_sequence_to_content,
     _is_base_message_params,
+    convert_message_content_to_message_param_content,
     convert_messages_to_message_params,
 )
 
@@ -120,7 +120,7 @@ def test_convert_message_sequence_to_content():
             audio=b"audio_bytes",
         ),
     ]
-    result = _convert_message_sequence_to_content(input_sequence)
+    result = convert_message_content_to_message_param_content(input_sequence)
     assert result == expected_output
 
 
@@ -150,15 +150,6 @@ def test_convert_messages_to_message_params_with_base_message_param():
     message_param = BaseMessageParam(role="user", content="Hello")
     result = convert_messages_to_message_params(message_param)
     assert result == [message_param]
-
-
-def test_convert_messages_to_message_params_with_base_message_param_list():
-    messages = [
-        BaseMessageParam(role="user", content="Hello"),
-        BaseMessageParam(role="assistant", content="Hi!"),
-    ]
-    result = convert_messages_to_message_params(messages)
-    assert result == messages
 
 
 def test_convert_messages_to_message_params_with_sequence():
